feat REPAUTO-6 REPAUTO-11: create content modal & clean up settings
