+++ src/frontend/components/Settings/Settings.jsx
@@ -1,34 +1,19 @@
 import React from "react";
-import GoalsOfTheDayList from "../GoalsOfTheDayList/GoalsOfTheDayList";
 
-import { 
-    Textfield,
-    Toggle,
-    Inline,
-    Button, 
-    Text, 
-    Box,
-    Strong,
-    TextArea,
-} from '@forge/react';
+import { Textfield } from '@forge/react';
 import BaseModal from "../BaseModal/BaseModal";
+import useSettings from "./useSettings";
 
 const Settings = ({ 
     isVisible, 
     setIsVisible,     
-    setSetting, 
-    slackEndpoint, 
-    sprintStatusEnabled, 
-    workProgressEnabled,
-    goalsOfTheDay,
-    newGoalOfTheDay,
-    setSlackEndpoint,
-    setSprintStatusEnabled,
-    setWorkProgressEnabled,
-    removeGoalOfTheDay,
-    addNewGoalOfTheDay,
-    setNewGoalOfTheDay 
 }) => {
+    const {
+        slackEndpoint,
+        setSlackEndpoint,
+        setSetting
+    } = useSettings();
+
     return (
         <BaseModal 
             isVisible={isVisible} 
@@ -45,61 +30,8 @@ const Settings = ({
                 placeholder='Slack Endpoint' 
                 value={slackEndpoint}
             />
-            <Box paddingBlock='space.300'>
-                <Inline alignBlock='center'>
-                    <Text>Goals of the Day:</Text>
-                </Inline>
-                <Box paddingBlock="space.300">
-                    {goalsOfTheDay.length === 0 ? (
-                        <Text>
-                            <Strong>No Goals yet...</Strong>
-                        </Text>
-                    ) : (
-                        <GoalsOfTheDayList 
-                            goalsOfTheDay={goalsOfTheDay} 
-                            handleRemoveClick={removeGoalOfTheDay} 
-                        />
-                    )}
-                </Box>
-                <Inline space='space.100' alignBlock='center'>
-                    <TextArea 
-                        width={600} 
-                        name='Add new Goal of the Day' 
-                        placeholder='Goal of the Day' 
-                        value={newGoalOfTheDay}
-                        onChange={(e) => setNewGoalOfTheDay(e.target.value)}
-                    />
-                    <Button appearance="primary" onClick={addNewGoalOfTheDay}>
-                        Add Goal
-                    </Button>
-                </Inline>
-            </Box>
-            <Box paddingBlock='space.300'>
-                <Inline alignBlock='center'>
-                    <Text>Sprint Status</Text>
-                    <Toggle 
-                        onChange={() => {
-                            setSprintStatusEnabled(!sprintStatusEnabled);
-                            setSetting({ sprintStatusEnabled: !sprintStatusEnabled });
-                        }}                        
-                        isChecked={sprintStatusEnabled}
-                        name='Sprint Status' 
-                    />
-                </Inline>
-                <Inline alignBlock='center'>
-                    <Text>Work Progress</Text>
-                    <Toggle 
-                        name='Work Progress' 
-                        onChange={() => {
-                            setWorkProgressEnabled(!workProgressEnabled);
-                            setSetting({ workProgressEnabled: !workProgressEnabled });
-                        }}         
-                        isChecked={workProgressEnabled}
-                    />
-                </Inline>
-            </Box>
-        </BaseModal>
+        </BaseModal>        
     );
-}
+};
 
 export default Settings;
