+++ src/frontend/components/Settings/useSettings.js
@@ -2,13 +2,8 @@ import { invoke } from '@forge/bridge';
 import { useEffect, useState } from 'react';
 import useIndex from '../../useIndex'
 
-const useSettingsModal = () => {
+const useSettings = () => {
     const [slackEndpoint, setSlackEndpoint] = useState('');
-    const [sprintStatusEnabled, setSprintStatusEnabled] = useState(false);
-    const [workProgressEnabled, setWorkProgressEnabled] = useState(false);
-    const [goalsOfTheDay, setGoalsOfTheDay] = useState([]);
-    const [newGoalOfTheDay, setNewGoalOfTheDay] = useState("");
-
     const {projectId} = useIndex();
 
     useEffect(() => {
@@ -16,9 +11,6 @@ const useSettingsModal = () => {
             (async () => {
                 const data = await invoke('get-settings', { projectId });
                 setSlackEndpoint(data.slackEndpoint || '');
-                setSprintStatusEnabled(data.sprintStatusEnabled);
-                setWorkProgressEnabled(data.workProgressEnabled);
-                setGoalsOfTheDay(data.goalsOfTheDay ?? [])
             })();
         }
     }, [projectId]);
@@ -27,44 +19,11 @@ const useSettingsModal = () => {
         await invoke('set-setting', {projectId, settings})
     }
 
-    const removeGoalOfTheDay = (goalToRemove) => {
-        const updatedGoalsOfTheDay = goalsOfTheDay.filter(goal => goal !== goalToRemove);
-        setGoalsOfTheDay(updatedGoalsOfTheDay);
-        setSetting({ goalsOfTheDay: updatedGoalsOfTheDay });
-    };
-
-    const addNewGoalOfTheDay = () => {
-        if(!goalsOfTheDay || !newGoalOfTheDay || !newGoalOfTheDay.trim()) return
-
-        const updatedGoalsOfTheDay = [...goalsOfTheDay, newGoalOfTheDay];
-        setGoalsOfTheDay(updatedGoalsOfTheDay);
-        setSetting({ goalsOfTheDay: updatedGoalsOfTheDay });
-        setNewGoalOfTheDay(""); 
-       
-    };
-
-    const clearGoalsOfTheDay = () => {
-        setSetting({ goalsOfTheDay: [] });
-        setGoalsOfTheDay([])
-        setNewGoalOfTheDay('')   
-    }
-
     return {
         slackEndpoint,
-        sprintStatusEnabled,
-        workProgressEnabled,
-        goalsOfTheDay,
-        newGoalOfTheDay,
         setSlackEndpoint,
-        setSprintStatusEnabled,
-        setWorkProgressEnabled,
-        setGoalsOfTheDay,
-        setNewGoalOfTheDay,
         setSetting,
-        removeGoalOfTheDay,
-        addNewGoalOfTheDay,
-        clearGoalsOfTheDay
     };
 }
 
-export default useSettingsModal;
+export default useSettings;
