+++ src/frontend/components/index.js
@@ -1,1 +1,2 @@
 export { default as Settings } from './Settings/Settings'
+export { default as Content } from './Content/Content'
